+++ CHANGELOG.rst
@@ -21,6 +21,7 @@ Internal changes:
  - Clean GCC environment variables in test suite. (:issue:`493`)
  - Fix problems from deployment of release 5.0. (:issue:`494`)
  - Use yaxmldiff for XML diffing in tests. (:issue:`495`)
+ - Change test and reference for XML diff to be the same as for the unified diff. (:issue:`509`)
 
 5.0 (11 June 2021)
 ------------------